refactor: better DDD file structure approach for new feature
