+++ frontend/tests/unit/api/shared.ts
@@ -1,0 +1,4 @@
+import { fetchApi } from '../../../src/shared/api/api';
+import { ApiError, NetworkError } from '../../../src/shared/api/api-error';
+
+export { fetchApi, ApiError, NetworkError };
